requirements.txt
django==5.0
mssql-django==1.4
djangorestframework
django-cors-headers
coverage